fraud-detector-ui/requirements.txt
flask==1.1.2
kafka-python==2.0.1
redis==3.5.2
cassandra-driver==3.23.0
